Personalize app visuals, improve structure, and update dependencies.
