+++ country-explorer/src/App.css
@@ -14,13 +14,21 @@
 }
 
 .App-header {
-  background-color: #282c34;
-  min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   font-size: calc(10px + 2vmin);
+  color: white;
+  padding: 1rem 0;
+  margin-top: 2rem;
+}
+
+.App-header h1 {
+  background-color: rgba(0, 0, 0, 0.6);
+  padding: 0.5rem 1rem;
+  border-radius: 10px;
+  display: inline-block;
   color: white;
 }
