+++ country-explorer/src/components/countryCard/CountryCard.js
@@ -1,8 +1,9 @@
+import './CountryCard.css';
 import { Card, Button } from 'react-bootstrap';
 
 const CountryCard = ({ country, onSelect }) => {
   return (
-    <Card>
+    <Card className="country-card">
       <Card.Img variant="top" src={country.flags?.png} alt={`Flag of ${country.name.official}`} />
       <Card.Body>
         <Card.Title>{country.name.official}</Card.Title>
@@ -10,7 +11,7 @@ const CountryCard = ({ country, onSelect }) => {
           Region: {country.region} <br />
           Population: {country.population.toLocaleString()}
         </Card.Text>
-        <Button variant="primary" onClick={() => onSelect(country)}>
+        <Button className="view-details-button" onClick={() => onSelect(country)}>
           View Details
         </Button>
       </Card.Body>
